Autoscale group now works
Removed the creation policy and update policy need to figuire out what
that means.
Updated the max and min size to 1 and 2.
Collected the metrix on the Autoscalegroup so that we scale in/out based
on the group stats rather then the ebc.
Modify the cool down to take longer to scale down. (taking a 1 hour hit
for each test, best to make some use of the boxes ;)

Updated the alarms to be @40% , for the scale down and 70% scale up.
Finally figured out how to use the ssm, and set the right permissions.
This allows me to use the remote execution from the aws console.