AutoscalingGroup and CodeDeploy are not happy
Making the grace period fo the AutoscalingGroup a large number.
Switching the delployment tag from dev to Dev-Environment. We are now
using the autosacling group to deploy the application.

Rename the index.html that is created by the image to sampleindex.html.
Could be having some sort of locking thing going on.

Installing code deploy as soon as possible.